requirements.txt
appnope==0.1.3
asttokens==2.2.1
asyncio==3.4.3
attrs==23.1.0
backcall==0.2.0
black==23.7.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.18.0
flaky==3.7.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
idna==3.4
iniconfig==2.0.0
ipykernel==6.24.0
ipython==8.14.0
isort==5.12.0
janus==1.0.0
jedi==0.18.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nbclient==0.8.0
nbformat==5.9.1
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.9.1
pluggy==1.2.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-dependency==0.5.1
python-dateutil==2.8.2
pyzmq==25.1.0
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
rsa==4.9
six==1.16.0
stack-data==0.6.2
testbook==0.4.2
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==1.26.16
watchdog==3.0.0
wcwidth==0.2.6
